refactor: unify product models
