--- src/app/core/models/product/product-bundle.model.ts
@@ -1,5 +1,0 @@
-import { Product } from './product.model';
-
-export interface ProductBundle extends Product {
-  type: 'Bundle';
-}
