--- src/app/core/models/product/product-variation-master.model.ts
@@ -1,13 +1,0 @@
-import { Price } from 'ish-core/models/price/price.model';
-import { VariationAttribute } from 'ish-core/models/product-variation/variation-attribute.model';
-
-import { Product } from './product.model';
-
-export interface VariationProductMaster extends Product {
-  type: 'VariationProductMaster';
-  variationAttributeValues?: VariationAttribute[];
-  minListPrice?: Price;
-  minSalePrice?: Price;
-  maxListPrice?: Price;
-  maxSalePrice?: Price;
-}
