+++ src/app/core/models/product/product.model.ts
@@ -2,6 +2,7 @@ import { AttributeGroup } from 'ish-core/models/attribute-group/attribute-group.
 import { Attribute } from 'ish-core/models/attribute/attribute.model';
 import { Image } from 'ish-core/models/image/image.model';
 import { Price } from 'ish-core/models/price/price.model';
+import { VariationAttribute } from 'ish-core/models/product-variation/variation-attribute.model';
 import { SeoAttributes } from 'ish-core/models/seo-attributes/seo-attributes.model';
 
 export interface Product {
@@ -27,11 +28,40 @@ export interface Product {
   packingUnit: string;
 
   // properties added in model
-  type: 'Product' | 'VariationProduct' | 'VariationProductMaster' | 'Bundle' | 'RetailSet' | string;
+  type: string;
   promotionIds: string[];
   completenessLevel: number;
   failed: boolean;
   seoAttributes?: SeoAttributes;
 }
+
+export interface VariationProduct extends Product {
+  type: 'VariationProduct';
+  variableVariationAttributes?: VariationAttribute[];
+  productMasterSKU?: string;
+}
+
+export interface VariationProductMaster extends Product {
+  type: 'VariationProductMaster';
+  variationAttributeValues?: VariationAttribute[];
+  minListPrice?: Price;
+  minSalePrice?: Price;
+  maxListPrice?: Price;
+  maxSalePrice?: Price;
+}
+
+export interface ProductRetailSet extends Product {
+  type: 'RetailSet';
+  minListPrice: Price;
+  minSalePrice: Price;
+  summedUpListPrice: Price;
+  summedUpSalePrice: Price;
+}
+
+export interface ProductBundle extends Product {
+  type: 'Bundle';
+}
+
+export type AllProductTypes = Product | VariationProduct | VariationProductMaster | ProductBundle | ProductRetailSet;
 
 export * from './product.helper';
