+++ src/app/core/routing/product/product.route.spec.ts
@@ -7,8 +7,7 @@ import { Store, select } from '@ngrx/store';
 import { createCategoryView } from 'ish-core/models/category-view/category-view.model';
 import { Category } from 'ish-core/models/category/category.model';
 import { createProductView } from 'ish-core/models/product-view/product-view.model';
-import { VariationProduct } from 'ish-core/models/product/product-variation.model';
-import { Product } from 'ish-core/models/product/product.model';
+import { Product, VariationProduct } from 'ish-core/models/product/product.model';
 import { CoreStoreModule } from 'ish-core/store/core/core-store.module';
 import { selectRouter } from 'ish-core/store/core/router';
 import { categoryTree } from 'ish-core/utils/dev/test-data-utils';
