+++ src/app/shared/components/product/product-variation-select/product-variation-select.component.spec.ts
@@ -6,8 +6,7 @@ import { anything, capture, instance, mock, verify, when } from 'ts-mockito';
 
 import { ProductContextFacade } from 'ish-core/facades/product-context.facade';
 import { ProductView } from 'ish-core/models/product-view/product-view.model';
-import { VariationProductMaster } from 'ish-core/models/product/product-variation-master.model';
-import { VariationProduct } from 'ish-core/models/product/product-variation.model';
+import { VariationProduct, VariationProductMaster } from 'ish-core/models/product/product.model';
 import { findAllDataTestingIDs } from 'ish-core/utils/dev/html-query-utils';
 
 import { ProductVariationSelectComponent } from './product-variation-select.component';
